requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
pymongo==4.3.3
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.3.6